WIP: First steps loading multiple textures
