+++ src/js/models/floor.js
@@ -1,6 +1,6 @@
-import { vec3 } from 'gl-matrix'
 import TreeNode from './tree_node'
 import colors from '../constants/colors';
+import grassImage from '../../assets/maps/pasto.jpg'
 
 export default class Floor extends TreeNode {
     color() {
@@ -11,8 +11,11 @@ export default class Floor extends TreeNode {
 
         let pos = [];
         let normal = [];
+        let uv = [];
         let rows = 25;
         let cols = 25;
+
+        this.initTexture(grassImage);
 
         let n = [0, 1, 0];
 
@@ -26,6 +29,9 @@ export default class Floor extends TreeNode {
                 normal.push(n[0]);
                 normal.push(n[1]);
                 normal.push(n[2]);
+
+                //uv.push(j);
+                //uv.push(i);
             }
         }
 
@@ -37,11 +43,32 @@ export default class Floor extends TreeNode {
         gl.bindBuffer(gl.ARRAY_BUFFER, trianglesNormalBuffer);
         gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(normal), gl.STATIC_DRAW);
 
+        //let trianglesUvBuffer = gl.createBuffer();
+        //trianglesUvBuffer.number_points = uv.length;
+        //gl.bindBuffer(gl.ARRAY_BUFFER, trianglesUvBuffer);
+        //gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(uv), gl.STATIC_DRAW);
+
         return {
             vertexBuffer: trianglesVerticeBuffer,
             normalBuffer: trianglesNormalBuffer,
-            indexBuffer: this.buildIndexBuffer(rows, cols)
+            indexBuffer: this.buildIndexBuffer(rows, cols),
+            //uVBuffer: trianglesUvBuffer
         }
     }
+/*
+    draw(modelMatrix, viewMatrix, projMatrix) {
+        super.draw(modelMatrix, viewMatrix, projMatrix);
+        let trianglesUvBuffer = this.uVBuffer();
+        let vertexUvAttribute = gl.getAttribLocation(glProgram, "aVertexUv");
+        gl.enableVertexAttribArray(vertexUvAttribute);
+        gl.bindBuffer(gl.ARRAY_BUFFER, trianglesUvBuffer);
+        gl.vertexAttribPointer(vertexUvAttribute, 2, gl.FLOAT, false, 0, 0);
+
+        gl.activeTexture(gl.TEXTURE2);
+        gl.bindTexture(gl.TEXTURE_2D, this.texture);
+        gl.uniform1i(glProgram.samplerUniform, 0);
+                    
+        gl.drawArrays(gl.TRIANGLES, 0,trianglesUvBuffer.number_points);
+    }*/
 
 }
