+++ src/js/models/skybox.js
@@ -7,26 +7,6 @@ export default class Skybox extends Sphere {
         return colors.skyBlue;
     }
 
-    initTexture(filePath) {
-        this.texture = gl.createTexture();
-        this.texture.image = new Image();
-
-        let texture = this.texture;
-        this.texture.image.onload = function () {
-
-            gl.bindTexture(gl.TEXTURE_2D, texture); 						// activo la textura
-
-            gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, texture.image);	// cargo el bitmap en la GPU
-            gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);					// selecciono filtro de magnificacion
-            gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR_MIPMAP_NEAREST);	// selecciono filtro de minificacion
-
-            gl.generateMipmap(gl.TEXTURE_2D);		// genero los mipmaps
-            gl.bindTexture(gl.TEXTURE_2D, null);
-        }
-        this.texture.image.src = filePath;
-
-    }
-
     buildBuffers() {
         let pos = [];
         let normal = [];
@@ -56,11 +36,8 @@ export default class Skybox extends Sphere {
                 normal.push(-n[1]);
                 normal.push(-n[2]);
 
-                let uStep = alphaStep;
-                let vStep = (i + 17) / (rows - 1);
-
-                let u = uStep;
-                let v = vStep;
+                let u = alphaStep;
+                let v = (i + 17) / (rows - 1);
 
                 uv.push(u);
                 uv.push(v);
@@ -76,7 +53,7 @@ export default class Skybox extends Sphere {
         gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(normal), gl.STATIC_DRAW);
 
         let trianglesUvBuffer = gl.createBuffer();
-        trianglesUvBuffer.number_point = uv.length;
+        trianglesUvBuffer.number_points = uv.length;
         gl.bindBuffer(gl.ARRAY_BUFFER, trianglesUvBuffer);
         gl.bufferData(gl.ARRAY_BUFFER, new Float32Array(uv), gl.STATIC_DRAW);
 
@@ -86,5 +63,19 @@ export default class Skybox extends Sphere {
             indexBuffer: this.buildIndexBuffer(rows, cols),
             uVBuffer: trianglesUvBuffer
         }
+    }
+
+    draw(modelMatrix, viewMatrix, projMatrix) {
+        super.draw(modelMatrix, viewMatrix, projMatrix);
+        let trianglesUvBuffer = this.uVBuffer();
+        let vertexUvAttribute = gl.getAttribLocation(glProgram, "aVertexUv");
+        gl.enableVertexAttribArray(vertexUvAttribute);
+        gl.bindBuffer(gl.ARRAY_BUFFER, trianglesUvBuffer);
+        gl.vertexAttribPointer(vertexUvAttribute, 2, gl.FLOAT, false, 0, 0);
+
+        gl.activeTexture(gl.TEXTURE0);
+        gl.bindTexture(gl.TEXTURE_2D, this.texture);
+        gl.uniform1i(glProgram.samplerUniform, 0);
+        gl.drawArrays(gl.TRIANGLES, 0,trianglesUvBuffer.number_points);
     }
 }
