+++ src/shaders/fragment.glsl
@@ -17,5 +17,5 @@ void main(void) {
     
 	vec4 textureColor = texture2D(uSampler,vUv);
 	colorFinal = mix(colorFinal, textureColor.xyz, 0.8);
-	gl_FragColor = vec4(colorFinal, 1.0);
+	gl_FragColor = vec4(textureColor.xyz, 1.0);
 }
